# backend/internal/openapi/spec.yaml
openapi: 3.0.3
info:
  title: AI API Gateway
  description: |
    OpenAI-compatible gateway with tenant API keys, quota billing, redeem/invite flows, and admin APIs.
    Machine-readable spec is served at `GET /openapi.yaml`.
  version: 0.1.0
servers:
  - url: http://localhost:8080
    description: Local default (see `HTTP_ADDR`)
...
paths:
...
  /admin/channels:
    get:
      tags: [Admin]
      security: [{ bearerJWT: [] }]
      summary: List upstream channels
      responses:
        "200":
          description: Channel list (API key masked)
    post:
      tags: [Admin]
      security: [{ bearerJWT: [] }]
      summary: Create channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelPayload"
      responses:
        "201":
          description: Created
  /admin/channels/{id}:
    put:
      tags: [Admin]
      security: [{ bearerJWT: [] }]
      summary: Update channel
      parameters:
        - $ref: "#/components/parameters/ChannelId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelPayload"
      responses:
        "200":
          description: OK
    delete:
      tags: [Admin]
      security: [{ bearerJWT: [] }]
      summary: Delete channel
      parameters:
        - $ref: "#/components/parameters/ChannelId"
      responses:
        "200":
          description: OK
...
  /v1/models:
    get:
      tags: [Gateway]
      security: [{ gatewayApiKey: [] }]
      summary: List models (DB-backed by default)
      parameters:
        - in: query
          name: source
          schema: { type: string, enum: [upstream] }
          description: When upstream and API key configured, proxy env upstream
      responses:
        "200":
          description: OpenAI-style model list
        "401":
          description: Missing or invalid sk- key
  /v1/chat/completions:
    post:
      tags: [Gateway]
      security: [{ gatewayApiKey: [] }]
      summary: Chat completions (OpenAI-compatible body); may return 402 on insufficient quota
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: OpenAI chat completion request
      responses:
        "200":
          description: Completion or SSE stream
        "401":
          description: Invalid API key
        "402":
          description: Insufficient quota / unknown model pricing
        "429":
          description: Rate limited
        "502":
          description: Upstream unavailable after retries
...
components:
  parameters:
    ChannelId:
      in: path
      name: id
      required: true
      schema: { type: integer, format: int64 }
  securitySchemes:
    bearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: From `POST /auth/login` (`access_token`).
    gatewayApiKey:
      type: http
      scheme: bearer
      description: API key with prefix `sk-` (only for `/v1/*`).
  schemas:
...
    ChannelPayload:
      type: object
      required: [name, provider, base_url, api_key]
      properties:
        name: { type: string }
        provider: { type: string }
        base_url: { type: string }
        api_key: { type: string }
        models:
          type: array
          items: { type: string }
        model_mapping:
          type: object
          additionalProperties: { type: string }
        priority: { type: integer }
        weight: { type: integer }
        status: { type: integer }
        rate_limit: { type: integer }